mkqcdtbootimg: Implementation of Qualcomm mkbootimg
The Qualcomm boot.img format includes a set of device tree
blobs and a directory for these. From Qualcomm we get a tool
that generate the dtb.img and a patched up mkbootimg.
This tool combines the dtbTool and mkbootimg that can be
used with CUSTOM_MKBOOTIMG from the BoardConfig.mk of a device.