mkqcdtbootimg: Make it possible to specify qcdt version
Neither the qcdt format nor the bootloaders support mixing version 1 and
version 2 qcdt headers, so we need to be able to tell the tool what
format we expect if both versions are available in the given directory.

Therefor we add the --dt_version that will select what version will be
accepted into the boot.img.

Signed-off-by: Bjorn Andersson <bjorn.andersson@sonymobile.com>